always prefix json-ld context with json-ld and move event class to sub-namespace
