+++ includes/activity/class-activity.php
@@ -17,7 +17,7 @@ use Activitypub\Activity\Base_Object;
  * @see https://www.w3.org/TR/activitystreams-core/#intransitiveactivities
  */
 class Activity extends Base_Object {
-	const CONTEXT = array(
+	const JSON_LD_CONTEXT = array(
 		'https://www.w3.org/ns/activitystreams',
 	);
 
@@ -173,10 +173,10 @@ class Activity extends Base_Object {
 	 */
 	public function get_json_ld_context() {
 		if ( $this->object instanceof Base_Object ) {
-			// Without php 5.6 support this could be just: 'return  $this->object::CONTEXT;'
+			// Without php 5.6 support this could be just: 'return  $this->object::JSON_LD_CONTEXT;'
 			return call_user_func( array( get_class( $this->object ), 'CONTEXT' ) );
 		} else {
-			return self::CONTEXT;
+			return self::JSON_LD_CONTEXT;
 		}
 	}
 }
